requirements_lora.txt
absl-py==1.4.0
accelerate==0.12.0
addict==2.4.0
aenum==3.1.11
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
basicsr==1.4.2
beautifulsoup4==4.12.0
bitsandbytes==0.37.2
blendmodes==2022
boltons==23.0.0
cachetools==5.3.0
certifi
chardet==4.0.0
charset-normalizer==3.1.0
clean-fid==0.1.29
click==8.1.3
clip==0.2.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
deprecation==2.1.0
diffusers==0.14.0
easygui==0.98.3
einops==0.4.1
entrypoints==0.4
facexlib==0.2.5
fairscale==0.4.13
fastapi==0.90.1
ffmpy==0.3.0
filelock==3.10.4
filterpy==1.4.5
flatbuffers==23.3.3
font-roboto==0.0.1
fonts==0.0.3
fonttools==4.39.2
frozenlist==1.3.3
fsspec==2023.3.0
ftfy==6.1.1
future==0.18.3
gast==0.4.0
gdown==4.7.1
gfpgan==1.3.8
gitdb==4.0.10
GitPython==3.1.27
google-auth==2.16.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gradio==3.16.2
grpcio==1.51.3
h11==0.12.0
h5py==3.8.0
httpcore==0.15.0
httpx==0.23.3
huggingface-hub==0.13.3
humanfriendly==10.0
idna==2.10
imageio==2.26.1
importlib-metadata==6.1.0
importlib-resources==5.12.0
inflection==0.5.1
invisible-watermark==0.1.5
jax==0.4.6
Jinja2==3.1.2
joblib==1.2.0
jsonmerge==1.8.0
jsonschema==4.17.3
keras==2.12.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
kornia==0.6.7
lark==1.1.2
lazy_loader==0.2
libclang==16.0.0
Library==0.0.0
lightning-utilities==0.8.0
linkify-it-py==2.0.0
llvmlite==0.39.1
lmdb==1.4.0
lpips==0.1.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.0
numba==0.56.4
numpy==1.23.3
oauthlib==3.2.2
omegaconf==2.2.3
onnx==1.13.1
onnxruntime==1.14.1
open-clip-torch==2.16.0
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
orjson==3.8.8
packaging==23.0
pandas==1.5.3
piexif==1.1.3
Pillow==9.4.0
protobuf==3.20.3
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.17
pydantic==1.10.7
pyDeprecate==0.3.2
pydub==0.25.1
pyparsing==3.0.9
pyre-extensions==0.0.23
pyreadline3==3.4.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==1.7.6
pytz==2023.2
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
realesrgan==0.3.0
regex==2023.3.23
requests==2.25.1
requests-oauthlib==1.3.1
resize-right==0.0.2
rfc3986==1.5.0
rsa==4.9
safetensors==0.2.7
scikit-image==0.19.2
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.1.97
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
starlette==0.23.1
sympy==1.11.1
tb-nightly==2.13.0a20230317
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.3.21
timm==0.6.7
tk==0.1.0
tokenizers==0.13.2
toolz==0.12.0
torchdiffeq==0.2.3
torchmetrics==0.11.4
torchsde==0.2.5
tqdm==4.65.0
trampoline==0.1.2
transformers==4.25.1
typing-inspect==0.8.0
typing_extensions==4.5.0
uc-micro-py==1.0.1
urllib3==1.26.15
uvicorn==0.21.1
wcwidth==0.2.6
websockets==10.4
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.14.1
xformers==0.0.16
yapf==0.32.0
yarl==1.8.2
zipp==3.15.0







